some scrollwrite changes (#19)
This adds an option to the sound handling so that an in-mission looping sound is played for as long as the text is scrolling, and then stops when the text finishes scrolling.  Right now beep_5.wav is used as a placeholder sound effect; textdraw.wav could also be used, but it would have to be copied into the in-mission sound section in sounds.tbl.

Also adds an important detail to sexp help, removes trailing whitespace in the script, and fixes tab-indentation.

(Unfortunately the whitespace changes in `Scroll:Draw` make it appear that there are more changes than there really are.  The main non-whitespace changes are the addition of the `line.Handle` management and the removal of the `finalString` variable.)

This corresponds to the old PR 3:
https://github.com/AxemP/AxemFS2Scripts/pull/3