sync the "attack any" order
There is not a specific comment indicating Attack and Attack Any should be combined, so I'm guessing this was an oversight.

I've also added a check that the targets of orders are ships. Somehow a checkpoint in Colt's campaign kept triggering this on a target that was not a ship.

This corresponds to the old PR 7:
AxemP/AxemFS2Scripts#7