fix SaveLoad for an obscure edge case (#25)
If you have a ship that hasn't arrived yet, and that ship has a turret with a changed weapon on it, then restoring a checkpoint will cause a crash. This is because subsystems on parse objects use the weapon class directly, not weaponbanktype. This PR fixes the crash.

This corresponds to the old PR 6:
AxemP/AxemFS2Scripts#6