add option to keep HUD hidden when exiting visual novel scene
The visual novel script usually re-enables the HUD when a scene ends, but the KeepHUDHidden option now allows that to be skipped.

This also requires adding an extension to the validator to accept multiple data types.  However this extension is needed anyway as a bugfix, since the 'lockdown' action can accept a string as well as a number.