use isHiRes check to select the appropriate StandardHead
The standard talking head appears differently in the low-res and high-res HUDs.  Using the new `:isHiRes()` function, the script can choose the proper head to display.  The code is marked off with version-specific comments to avoid bumping the FSO build requirements.  For legacy builds, the StandardHead will default to the high-res talking head used previously.  (And of course modders can change this by editing the script if desired.)