call setClip() when calling setScreenScale()
In most places in the FreeSpace code, and always when displaying text, a call to set the screen scale is always followed by a call to set the clipping region.  So, make sure the same is done in all relevant scripts in this repository.  Without this fix, text can be cut off at high resolutions because the engine is using a clipping region smaller than the drawing area.