fix deprecation in radaricon script
The `self.Config.AtLeast_20_1_0_RC1` script variable is not the best way to accommodate version changes, as this relies on proper setting of the value in the config file, while version-specific commenting can be built into the script itself.  So, replace the code with version-specific commenting.  This also has the consequence of fixing the deprecation of `gr.drawMonochromeImage()` in recent builds.